Move allowed block filter to a util file
