+++ packages/block-editor/src/components/inserter/block-types-tab.native.js
@@ -10,21 +10,17 @@ import BlockTypesList from '../block-types-list';
 import useClipboardBlock from './hooks/use-clipboard-block';
 import { store as blockEditorStore } from '../../store';
 import useBlockTypeImpressions from './hooks/use-block-type-impressions';
+import { filterInserterItems } from '.utils';
 
-function BlockTypesTab( {
-	onSelect,
-	rootClientId,
-	listProps,
-	allowedBlockFilter = () => true,
-} ) {
+function BlockTypesTab( { onSelect, rootClientId, listProps } ) {
 	const clipboardBlock = useClipboardBlock( rootClientId );
 
 	const { blockTypes } = useSelect(
 		( select ) => {
 			const { getInserterItems } = select( blockEditorStore );
-
-			const allItems = getInserterItems( rootClientId );
-			const blockItems = allItems.filter( allowedBlockFilter );
+			const blockItems = filterInserterItems(
+				getInserterItems( rootClientId )
+			);
 
 			return {
 				blockTypes: clipboardBlock
