+++ packages/block-editor/src/components/inserter/reusable-blocks-tab.native.js
@@ -1,7 +1,6 @@
 /**
  * WordPress dependencies
  */
-import { useCallback } from '@wordpress/element';
 import { useSelect } from '@wordpress/data';
 
 /**
@@ -9,24 +8,17 @@ import { useSelect } from '@wordpress/data';
  */
 import BlockTypesList from '../block-types-list';
 import { store as blockEditorStore } from '../../store';
+import { filterInserterItems } from './utils';
 
-function ReusableBlocksTab( {
-	onSelect,
-	rootClientId,
-	listProps,
-	allowedBlockFilter = () => true,
-} ) {
-	const blockFilter = useCallback( ( block ) =>
-		allowedBlockFilter( block, { onlyReusable: true } )
-	);
-
+function ReusableBlocksTab( { onSelect, rootClientId, listProps } ) {
 	const { items } = useSelect(
 		( select ) => {
 			const { getInserterItems } = select( blockEditorStore );
 			const allItems = getInserterItems( rootClientId );
-			const reusableBlockItems = allItems.filter( blockFilter );
 
-			return { items: reusableBlockItems };
+			return {
+				items: filterInserterItems( allItems, { onlyReusable: true } ),
+			};
 		},
 		[ rootClientId ]
 	);
