+++ packages/block-editor/src/components/inserter/utils.native.js
@@ -1,0 +1,29 @@
+const REUSABLE_BLOCKS_CATEGORY = 'reusable';
+const ALLOWED_EMBED_VARIATIONS = [ 'core/embed' ];
+
+export function blockAllowed( block, { onlyReusable, allowReusable } ) {
+	const { id, category } = block;
+	const isReusable = category === REUSABLE_BLOCKS_CATEGORY;
+
+	if ( onlyReusable ) {
+		return isReusable;
+	}
+
+	if ( isReusable ) {
+		return allowReusable;
+	}
+	// We don't want to show all possible embed variations
+	// as different blocks in the inserter. We'll only show a
+	// few popular ones.
+	return category !== 'embed' || ALLOWED_EMBED_VARIATIONS.includes( id );
+}
+
+export function filterInserterItems(
+	items,
+	{ onlyReusable = false, allowReusable = false }
+) {
+	const allowBlock = ( block ) =>
+		blockAllowed( block, { onlyReusable, allowReusable } );
+
+	return items.filter( allowBlock );
+}
